Create redirect when an invalid page is open, removed old achievement
