+++ src/app/(dashboard)/teacher/achievement/components/achievementForm.tsx
@@ -1,143 +1,147 @@
-"use client"
-import { createUser, updateUser } from '@/actions/manageUserAction';
-import styles from '../styles/acheivement.module.scss';
-import { FormEvent, useState, useEffect } from 'react';
-import { useRouter } from "next/navigation";
+"use client";
+import { useState } from "react";
+import styles from "../styles/achievement.module.scss";
 
-interface UserFormProps {
-  closeModal: () => void;
-  initialData?: {
-    id?: number; // Ensure id is a number
-    name: string;
-    email: string;
-    role: string;
-  };
+interface AchievementData {
+  title: string;
+  category: string;
+  level: string;
+  certificate: File | null;
+  date: string;
+  description: string;
 }
 
-export default function UserForm({ closeModal, initialData }: UserFormProps) {
-  const [user, setUser] = useState({
-    name: '',
-    email: '',
-    role: 'teacher'
+interface AchievementFormProps {
+  onSubmit: (data: AchievementData) => void;
+  onClose: () => void;
+}
+
+export default function AchievementForm({ onSubmit, onClose }: AchievementFormProps) {
+  const [formData, setFormData] = useState<AchievementData>({
+    title: "",
+    category: "",
+    level: "",
+    certificate: null,
+    date: "",
+    description: "",
   });
 
-  const [isEditingPassword, setIsEditingPassword] = useState(false);
-  const [password, setPassword] = useState('');
-  const router = useRouter();
+  const [filePreview, setFilePreview] = useState<string | null>(null); // State for the file preview
 
-  useEffect(() => {
-    if (initialData) {
-      setUser(initialData);
-    }
-  }, [initialData]);
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
-    const { name, value } = event.target;
-    setUser((prevUser) => ({ ...prevUser, [name]: value }));
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
+    const { name, value } = e.target;
+    setFormData((prev) => ({ ...prev, [name]: value }));
   };
 
-  const handlePasswordChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setPassword(event.target.value);
-  };
+  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0] || null;
+    setFormData((prev) => ({ ...prev, certificate: file }));
 
-  const handleEditPassword = () => {
-    setIsEditingPassword(true);
-  };
-
-  async function handleSubmit(event: FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    if (initialData && typeof initialData.id === 'number') {
-      await updateUser(initialData.id, new FormData(event.target as HTMLFormElement));
+    if (file) {
+      // Generate a preview URL if the file is an image
+      if (file.type.startsWith("image/")) {
+        setFilePreview(URL.createObjectURL(file));
+      } else if (file.type === "application/pdf") {
+        // For PDFs, you can show a generic icon or thumbnail (you can replace with an icon later)
+        setFilePreview("pdf-icon.png"); // Replace with an actual PDF icon image if needed
+      } else {
+        setFilePreview(null); // For unsupported file types, clear the preview
+      }
     } else {
-      await createUser(new FormData(event.target as HTMLFormElement));
+      setFilePreview(null); // Clear the preview if no file is selected
     }
-    closeModal();
-    router.refresh();
-  }
+  };
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    onSubmit(formData);
+  };
 
   return (
-    <div className={styles.userFormSection}>
-      <div className={styles.modalHeader}>
-        <h2>{initialData ? 'Edit User' : 'Add User'}</h2>
-        <button type="button" className={styles.closeButton} onClick={closeModal} aria-label="Close modal">
-          &times;
-        </button>
-      </div>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="name">Name</label>
+    <form onSubmit={handleSubmit} className={styles.form}>
+      <h2 style={{ fontWeight: "bold" }}>Add New Achievement</h2>
+      <div className={styles.formGroup}>
+        <label htmlFor="title">Title</label>
         <input
           type="text"
-          id="name"
-          name="name"
-          value={user.name}
+          id="title"
+          name="title"
+          value={formData.title}
           onChange={handleChange}
-          className={styles.inputField}
           required
-          maxLength={100}
         />
-        
-        <label htmlFor="email">Email</label>
+      </div>
+      <div className={styles.formGroup}>
+        <label htmlFor="category">Category</label>
         <input
-          type="email"
-          id="email"
-          name="email"
-          value={user.email}
+          type="text"
+          id="category"
+          name="category"
+          value={formData.category}
           onChange={handleChange}
-          className={styles.inputField}
           required
         />
-        
-        <label htmlFor="role">Role</label>
-        <select
-          id="role"
-          name="role"
-          value={user.role}
+      </div>
+      <div className={styles.formGroup}>
+        <label htmlFor="level">Level</label>
+        <input
+          type="text"
+          id="level"
+          name="level"
+          value={formData.level}
           onChange={handleChange}
-          className={styles.inputField}
           required
-        >
-          <option value="Teacher">Teacher</option>
-          <option value="Admin">Admin</option>
-        </select>
-        {!initialData ? (
-          <div className={styles.formGroup}>
-            <label htmlFor="password">Password</label>
-            <input
-              type="password"
-              id="password"
-              name="password"
-              value={password}
-              onChange={handlePasswordChange}
-              className={styles.inputField}
-              required
-              minLength={8}
-            />
-          </div>
-        ) : (
-          <>
-            {isEditingPassword ? (
-              <div className={styles.formGroup}>
-                <label htmlFor="password">New Password</label>
-                <input
-                  type="password"
-                  id="password"
-                  name="password"
-                  value={password}
-                  onChange={handlePasswordChange}
-                  className={styles.inputField}
-                  required
-                  minLength={8}
-                />
-              </div>
+        />
+      </div>
+      <div className={styles.formGroup}>
+        <label htmlFor="certificate">Certificate</label>
+        <input
+          type="file"
+          id="certificate"
+          name="certificate"
+          onChange={handleFileChange}
+        />
+        {/* File Preview Section */}
+        {filePreview && (
+          <div className={styles.filePreview}>
+            {filePreview.includes("pdf") ? (
+              <img src="/path-to-pdf-icon.png" alt="PDF Preview" width={50} height={50} /> // Placeholder PDF icon
             ) : (
-              <button type="button" onClick={handleEditPassword} className={"btn btn-primary"}>
-                Edit Password
-              </button>
+              <img src={filePreview} alt="File Preview" style={{ maxWidth: "200px", maxHeight: "200px" }} />
             )}
-          </>
+          </div>
         )}
-        <button type="submit" className={styles.submitButton}>{initialData ? 'Save Changes' : 'Add User'}</button>
-      </form>
-    </div>
+      </div>
+      <div className={styles.formGroup}>
+        <label htmlFor="date">Date</label>
+        <input
+          type="date"
+          id="date"
+          name="date"
+          value={formData.date}
+          onChange={handleChange}
+          required
+        />
+      </div>
+      <div className={styles.formGroup}>
+        <label htmlFor="description">Description</label>
+        <textarea
+          id="description"
+          name="description"
+          value={formData.description}
+          onChange={handleChange}
+          rows={4}
+          required
+        ></textarea>
+      </div>
+      <div className={styles.actions}>
+        <button type="submit" className={styles.submitButton}>
+          Submit
+        </button>
+        <button type="button" className={styles.cancelButton} onClick={onClose}>
+          Cancel
+        </button>
+      </div>
+    </form>
   );
 }
