+++ src/app/(dashboard)/teacher/achievement/components/achievementList.tsx
@@ -1,35 +1,36 @@
 import prisma from "@/lib/db";
-import styles from '../styles/achievement.module.scss';
-import PaginationComponent from './Pagination';
+import styles from "../styles/achievement.module.scss";
+import PaginationComponent from "./Pagination";
 
-interface UserListProps {
+interface AchievementListProps {
   query: string;
   currentPage: number;
 }
 
 const ITEMS_PER_PAGE = 10;
 
-export default async function AchievementList({ query, currentPage }: UserListProps) {
+export default async function AchievementList({ query, currentPage }: AchievementListProps) {
   const offset = (currentPage - 1) * ITEMS_PER_PAGE;
 
-  const achievement = await prisma.achievementdata.findMany({
+  // Fetching achievements based on search query
+  const achievements = await prisma.achievementdata.findMany({
     where: {
       title: {
         contains: query,
-        mode: 'insensitive', // Case-insensitive search
+        mode: "insensitive", // Case-insensitive search
       },
     },
     skip: offset,
     take: ITEMS_PER_PAGE,
   });
 
-  const totalAchievement = await prisma.user.count();
-
-  const totalPages = Math.ceil(totalAchievement / ITEMS_PER_PAGE);
+  // Getting the total count for pagination
+  const totalAchievements = await prisma.achievementdata.count();
+  const totalPages = Math.ceil(totalAchievements / ITEMS_PER_PAGE);
 
   return (
-    <div className={styles.userListSection}>
-      <table className={styles.userTable}>
+    <div className={styles.achievementListSection}>
+      <table className={styles.achievementTable}>
         <thead>
           <tr>
             <th>Title</th>
@@ -39,17 +40,16 @@ export default async function AchievementList({ query, currentPage }: UserListPr
           </tr>
         </thead>
         <tbody>
-          {achievement.map((achievement) => (
+          {achievements.map((achievement) => (
             <tr key={achievement.id}>
               <td>{achievement.title}</td>
               <td>{achievement.category}</td>
               <td>{achievement.level}</td>
               <td>
                 <div className={styles.actionButtons}>
-                  <div className={styles.editButton}>
-                  </div>
-                  <div className={styles.deleteButton}><area shape="" coords="" href="" alt="" />
-                  </div>
+                  <button className={styles.viewButton}>View</button>
+                  <button className={styles.editButton}>Edit</button>
+                  <button className={styles.deleteButton}>Delete</button>
                 </div>
               </td>
             </tr>
