--- src/app/(dashboard)/teacher/achievement2/styles/achievement.module.scss
@@ -1,402 +1,0 @@
-/* Page Container */
-.achievementPage {
-  max-width: 100%;
-  margin: 0 auto;
-  padding: 20px;
-  font-family: "Arial", sans-serif;
-  border-radius: 10px;
-  background-color: #f8f8ff;
-}
-
-.pageContainer {
-  display: flex;
-  flex-direction: column;
-  height: 90vh;
-  padding: 20px;
-}
-
-/* Top Section */
-.topSection {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 20px;
-  border-bottom: 2px solid #ddd;
-  padding-bottom: 15px;
-}
-
-.searchAndActions {
-  display: flex;
-  align-items: center;
-  gap: 20px;
-  width: 100%;
-}
-
-.searchInput {
-  padding: 8px 12px;
-  font-size: 0.9rem;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  outline: none;
-  transition: border-color 0.3s ease;
-  width: 100%;
-  max-width: 250px;
-}
-
-.importButton,
-.addButton {
-  padding: 6px 12px;
-  background-color: #623988;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 0.9rem;
-  transition: background-color 0.3s ease;
-}
-
-.importButton:hover,
-.addButton:hover {
-  background-color: #42245e;
-}
-
-.sortButton {
-  display: flex;
-  align-items: center;
-  gap: 5px;
-  background-color: #FF8900;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  padding: 6px 12px;
-  cursor: pointer;
-  font-size: 0.9rem;
-  transition: background-color 0.3s ease;
-
-  &:hover {
-    background-color: #e6951d;
-  }
-}
-
-/* Achievement List Styles */
-.achievementListSection {
-  margin-top: 20px;
-  padding: 20px;
-  background-color: #f9f9f9;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-}
-
-.achievementTable {
-  width: 100%;
-  border-collapse: collapse;
-  background-color: #fff;
-  border-radius: 8px;
-  overflow: hidden;
-}
-
-.achievementTable th,
-.achievementTable td {
-  padding: 10px 15px;
-  text-align: left;
-  border-bottom: 1px solid #ddd;
-}
-
-.achievementTable th {
-  background-color: #623988;
-  color: white;
-  font-size: 1rem;
-}
-
-.achievementTable td {
-  font-size: 0.9rem;
-  color: #555;
-}
-
-.achievementTable tr:hover {
-  background-color: #f1f1f1;
-}
-
-/* Action Buttons */
-.actionButtons {
-  display: flex;
-  gap: 5px;
-
-  .viewButton{
-    background-color: #4CBD00;
-    color: white;
-
-    &:hover {
-      background-color: #3a9c00;
-    }
-  }
-
-  .editButton {
-    background-color: #007bff;
-    color: white;
-
-    &:hover {
-      background-color: #0056b3;
-    }
-  }
-
-  .deleteButton {
-    background-color: #dc3545;
-    color: white;
-
-    &:hover {
-      background-color: #bd1c2d;
-    }
-  }
-}
-
-/* Modal Styles */
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.modalContent {
-  background-color: white;
-  padding: 20px;
-  border-radius: 8px;
-  width: 90%;
-  max-width: 600px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.2);
-}
-
-/* Pagination */
-.pagination {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 20px;
-
-  select {
-    padding: 6px;
-    font-size: 0.9rem;
-    border-radius: 5px;
-    border: 1px solid #ccc;
-    transition: border-color 0.3s ease;
-
-    &:focus {
-      border-color: #007bff;
-    }
-  }
-
-  button {
-    background-color: #623988;
-    padding: 6px 12px;
-    border: none;
-    color: white;
-    border-radius: 5px;
-    font-size: 0.9rem;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-
-    &:hover {
-      background-color: #42245e;
-    }
-  }
-}
-
-/* Form Styling */
-.form {
-  background-color: #ffffff;
-  padding: 20px;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  width: 100%;
-  max-width: 500px;
-  margin: 20px auto;
-  font-family: "Arial", sans-serif;
-  max-height: 600px; /* Set a max-height for the form to make it scrollable */
-  overflow-y: auto; /* Enable vertical scrolling when content exceeds height */
-}
-
-.form h2 {
-  text-align: center;
-  color: #623988;
-  font-size: 1.5rem;
-  margin-bottom: 20px;
-}
-
-/* Form Group */
-.form .formGroup {
-  margin-bottom: 15px;
-}
-
-.form label {
-  display: block;
-  font-size: 1rem;
-  font-weight: 600;
-  color: #555;
-  margin-bottom: 5px;
-}
-
-.form input,
-.form textarea {
-  width: 100%;
-  padding: 10px;
-  font-size: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  outline: none;
-  transition: border-color 0.3s ease;
-}
-
-.form input[type="file"] {
-  padding: 5px;
-  border: 1px solid #ccc;
-}
-
-.form input:focus,
-.form textarea:focus {
-  border-color: #623988;
-}
-
-/* Textarea Specific */
-.form textarea {
-  resize: vertical;
-  min-height: 100px;
-  font-family: "Arial", sans-serif;
-}
-
-/* Actions */
-.form .actions {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 20px;
-}
-
-.form .submitButton,
-.form .cancelButton {
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.form .submitButton {
-  background-color: #623988;
-  color: white;
-}
-
-.form .submitButton:hover {
-  background-color: #42245e;
-}
-
-.form .cancelButton {
-  background-color: #f0f0f0;
-  color: #333;
-  border: 1px solid #ccc;
-}
-
-.form .cancelButton:hover {
-  background-color: #ddd;
-}
-
-/* Responsive Design */
-@media (max-width: 600px) {
-  .form {
-    padding: 15px;
-    width: 90%;
-  }
-  .form h2 {
-    font-size: 1.25rem;
-  }
-  .form .submitButton,
-  .form .cancelButton {
-    padding: 8px 15px;
-    font-size: 0.9rem;
-  }
-}
-
-/* In your SCSS file, add styles for the file preview */
-.filePreview {
-  margin-top: 10px;
-}
-
-.filePreview img {
-  max-width: 100%;
-  max-height: 200px;
-  border-radius: 5px;
-}
-
-
-
-.fileDropArea {
-  border: 2px dashed #ccc;
-  padding: 20px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.fileDropArea.dragging {
-  background-color: #f0f8ff;
-}
-
-.chooseFileLabel {
-  display: inline-block;
-  margin-top: 10px;
-}
-
-.chooseFileButton {
-  background-color: #4caf50;
-  color: white;
-  padding: 10px 20px;
-  font-size: 16px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.chooseFileButton:hover {
-  background-color: #45a049;
-}
-
-.fileInput {
-  display: none; /* Hide default file input */
-}
-
-.submitButton,
-.cancelButton {
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.submitButton {
-  background-color: #4caf50;
-  color: white;
-}
-
-.submitButton:hover {
-  background-color: #45a049;
-}
-
-.cancelButton {
-  background-color: #f44336;
-  color: white;
-}
-
-.cancelButton:hover {
-  background-color: #e53935;
-}
-
-.actions {
-  margin-top: 20px;
-}
-
-
