+++ src/app/(dashboard)/teacher/layout.tsx
@@ -1,5 +1,4 @@
 import type { Metadata } from "next";
-import SidebarAdmin from '@/components/sidebar/sidebarAdmin';
 import SidebarTeacher from '@/components/sidebar/sidebarTeacher' // Ensure the correct path to your SidebarAdmin component
 import Header from '@/components/header/header'; 
 // import AdminGuard from "./components/AdminGuard";
